# src/HouseholdLedger.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HouseholdLedger.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pay-schedules": {
      "get": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Lists recurring income pay schedules.",
        "responses": {
          "200": {
            "description": "The pay schedules.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PayScheduleResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Creates one recurring income pay schedule.",
        "requestBody": {
          "description": "The request cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayScheduleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayScheduleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pay-schedules/{scheduleId}": {
      "get": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Gets one recurring income pay schedule.",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "The schedule identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayScheduleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Revises the future settings for one recurring income pay schedule.",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "The schedule identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The request cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisePayScheduleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisePayScheduleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RevisePayScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pay-schedules/receipts": {
      "get": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Lists user-confirmed income receipts in an inclusive calendar range.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "The inclusive first pay date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "The inclusive final pay date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeReceiptResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pay-schedules/{scheduleId}/pause": {
      "post": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Pauses one recurring income pay schedule.",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "The schedule identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pay-schedules/{scheduleId}/resume": {
      "post": {
        "tags": [
          "PaySchedules"
        ],
        "summary": "Resumes future income expectation suggestions without restoring missed dates.",
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "The schedule identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The request cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumePayScheduleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumePayScheduleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResumePayScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePayScheduleRequest": {
        "required": [
          "name",
          "firstPayDate",
          "cadence",
          "netIncome",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The user-visible schedule name."
          },
          "firstPayDate": {
            "type": "string",
            "description": "The first pay date.",
            "format": "date"
          },
          "cadence": {
            "description": "The schedule recurrence cadence.",
            "$ref": "#/components/schemas/PayPeriodCadence"
          },
          "netIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive net income for each receipt.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationRequest"
            },
            "description": "The account allocations for each receipt."
          },
          "secondMonthlyPayDay": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "The second pay day for a semimonthly schedule.",
            "format": "int32"
          }
        },
        "description": "Requests creation of a recurring income pay schedule."
      },
...
      "IncomeAllocationRequest": {
        "required": [
          "accountId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The destination account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive amount allocated to the account.",
            "format": "double"
          }
        },
        "description": "Requests the allocation of one income receipt amount to an account."
      },
      "IncomeAllocationResponse": {
        "required": [
          "accountId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The destination account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The allocated amount.",
            "format": "double"
          }
        },
        "description": "Describes one income allocation to an account."
      },
      "IncomeReceiptResponse": {
        "required": [
          "id",
          "scheduleId",
          "payDate",
          "netIncome",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The receipt identifier.",
            "format": "uuid"
          },
          "scheduleId": {
            "type": [
              "null",
              "string"
            ],
            "description": "The optional expectation schedule identifier.",
            "format": "uuid"
          },
          "payDate": {
            "type": "string",
            "description": "The receipt pay date.",
            "format": "date"
          },
          "netIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The received net income.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationResponse"
            },
            "description": "The immutable account allocation snapshot."
          }
        },
        "description": "Describes a confirmed income receipt."
      },
...
      "PayPeriodCadence": {
        "type": "integer",
        "description": "Defines the recurrence cadence for a pay schedule."
      },
      "PayScheduleResponse": {
        "required": [
          "id",
          "name",
          "firstPayDate",
          "cadence",
          "netIncome",
          "allocations",
          "secondMonthlyPayDay",
          "isPaused"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The schedule identifier.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "The user-visible schedule name."
          },
          "firstPayDate": {
            "type": "string",
            "description": "The first pay date.",
            "format": "date"
          },
          "cadence": {
            "description": "The recurrence cadence.",
            "$ref": "#/components/schemas/PayPeriodCadence"
          },
          "netIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The net income for each receipt.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationResponse"
            },
            "description": "The account allocations for each receipt."
          },
          "secondMonthlyPayDay": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "The second pay day for a semimonthly schedule.",
            "format": "int32"
          },
          "isPaused": {
            "type": "boolean",
            "description": "Whether future schedule expectations are paused."
          }
        },
        "description": "Describes a recurring income pay schedule."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ResumePayScheduleRequest": {
        "required": [
          "resumeDate"
        ],
        "type": "object",
        "properties": {
          "resumeDate": {
            "type": "string",
            "description": "The date from which future schedule expectations are eligible.",
            "format": "date"
          }
        },
        "description": "Requests resumption of a pay schedule from a date."
      },
      "RevisePayScheduleRequest": {
        "required": [
          "name",
          "firstPayDate",
          "cadence",
          "netIncome",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The replacement schedule name."
          },
          "firstPayDate": {
            "type": "string",
            "description": "The first pay date using the replacement settings.",
            "format": "date"
          },
          "cadence": {
            "description": "The replacement recurrence cadence.",
            "$ref": "#/components/schemas/PayPeriodCadence"
          },
          "netIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The replacement net income for each receipt.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationRequest"
            },
            "description": "The replacement account allocations for each receipt."
          },
          "secondMonthlyPayDay": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "The replacement second pay day for a semimonthly schedule.",
            "format": "int32"
          }
        },
        "description": "Requests replacement of a pay schedule's future settings."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}